improved impersonation
